Fix amount parsing in URIs

+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/utils/PaymentAddressParser.kt
@@ -49,7 +49,7 @@ class PaymentAddressParser(private val removeScheme: Boolean) {
             if (parts.size == 2) {
                 when (parts[0]) {
                     parameterGas -> gas = parts[1].toLong()
-                    parameterValue -> value = parts[1].toBigInteger()
+                    parameterValue -> value = parts[1].toBigDecimal().toBigInteger()
                     parameterGasLimit-> gas = parts[1].toLong()
                     parameterGasPrice -> gasPrice = parts[1].toLong()
                     else -> parameters[parts[0]] = URLDecoder.decode(parts[1], "UTF-8")
